working on the db backup script

+++ tools/backup-webdb.sh
@@ -1,12 +1,19 @@
-#!/bin/sh
-
+#!/bin/bash
+dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+base=`dirname $dir`
 fname=`date +"%Y%m%d-%H%M"`.db.lzma
-db=4b/main.db
+db=$base/db/main.db
 backup=/tmp/backup.lzma
-source ../misc/backup-secrets
+if [ ! -e $base/misc/backup-secrets.ini ]; then
+  echo "Erorr, $base/misc/backup-secrets.ini needs to exist."
+  exit -1
+fi
 
-[ -e $backup ] && rm $backup 
-lzma -c 9 $db $backup
+source $base/misc/backup-secrets.ini
+
+rm -f $backup 
+lzma -9 -c $db > $backup
+
 if [ -e $backup ]; then
   scp -P $port -i $private_key $backup $username@$host:$path/$fname
   rm $backup
